requirements.txt
python-dotenv==1.0.0
structlog==23.3.0
GitPython==3.1.40 